requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
conda==4.7.10
conda-package-handling==1.3.11
cryptography==2.7
Django==2.2.6
facebook-sdk==3.1.0
googlemaps==4.2.0
gunicorn==20.0.4
idna==2.8
libarchive-c==2.8
oauthlib==3.1.0
Pillow==7.0.0
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel-yaml==0.15.46
six==1.12.0
sqlparse==0.3.0
tqdm==4.32.1
tweepy==3.8.0
urllib3==1.24.2
